refactor SlowQueriesTable
