+++ ui/lib/apps/SlowQuery/index.tsx
@@ -2,7 +2,6 @@ import React from 'react'
 import { Root } from '@lib/components'
 import { HashRouter as Router, Route, Routes } from 'react-router-dom'
 import { List, Detail } from './pages'
-import useSlowQueryTableController from './utils/useSlowQueryTableController'
 
 export default function () {
   return (
@@ -19,4 +18,5 @@ export default function () {
 
 export * from './components'
 export * from './pages'
-export { useSlowQueryTableController }
+export * from './utils/useSlowQueryTableController'
+export { default as useSlowQueryTableController } from './utils/useSlowQueryTableController'
